feat: provide user interface for specifying replicates and reporting groups
replicates allow multiple estimates of precision and recall for the same conditions but run with
different preps. not sure how useful this will be, but it's a component of the existing internal
runs. the comparisons can now also be specified as groups that will be included in particular
reports. this information is accepted in the manifests, but the code doesn't actually do anything
with it yet.