feat: introduce the distinction between stratification regions and confident regions
the previously included bedfiles needed to be provided to the stratification functionality of
hap.py, whereas a different bedfile or set of bedfiles was required to specify so-called confident
regions. confident regions are now specified in config, and all combinations of confident regions
and comparisons are run as separate reports.