feat: track individual files within stratification regions with http remotes
an end user reported issues with lftp being unable to connect behind sufficient stringent firewalls.
the tracking in the workflow was lazy in any case. I've updated things to checkpoint on the
stratification file linker correctly and track stratification set provenance individually. one minor
issue here is that the full set of stratifications is now tracked for all happy comparisons; this is
overly conservative in some instances, and will create a fairly heavy spurious rerun burden .